feature/collision: Implement collision and pathfinder from blurite.

+++ game-server/common/src/main/kotlin/com/osrs/common/buffer/BytePacketBuilder.kt
@@ -48,7 +48,7 @@ fun BytePacketBuilder.writeShortAdd(value: Short) {
     writeByteAdd(value.toByte())
 }
 
-fun BytePacketBuilder.writeShortLittleEndianAdd(value: Short) {
+fun BytePacketBuilder.writeShortLittleEndianAdd(value: Int) {
     writeByteAdd(value.toByte())
     writeByte((value.toInt() shr 8).toByte())
 }
